feat(date range): add date range  picker to product-edit page and factor shared utility functions
